RF/EN: Added additional gltools functions for mesh generation. Split the huge file into a package.
