+++ psychopy/tools/gltools/_light.py
@@ -1,0 +1,122 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+"""OpenGL for working with lighting.
+
+These functions and classes are used for managing lighting using OpenGL's
+immediate mode pipeline. Note that defining lights this way have been
+depreciated and these routines might not work as intended on some platforms.
+
+"""
+
+# Part of the PsychoPy library
+# Copyright (C) 2002-2018 Jonathan Peirce (C) 2019-2020 Open Science Tools Ltd.
+# Distributed under the terms of the GNU General Public License (GPL).
+
+__all__ = [
+    'Light',
+    'createLight',
+    'useLights',
+    'setAmbientLight'
+]
+
+from collections import namedtuple
+import pyglet.gl as GL
+from ._misc import getIntegerv
+
+# -------------------------
+# Lighting Helper Functions
+# -------------------------
+
+Light = namedtuple('Light', ['params', 'userData'])
+
+
+def createLight(params=()):
+    """Create a point light source.
+
+    """
+    # setup light mode/value slots
+    lightDesc = Light({mode: None for mode in (
+        GL.GL_AMBIENT,
+        GL.GL_DIFFUSE,
+        GL.GL_SPECULAR,
+        GL.GL_POSITION,
+        GL.GL_SPOT_CUTOFF,
+        GL.GL_SPOT_DIRECTION,
+        GL.GL_SPOT_EXPONENT,
+        GL.GL_CONSTANT_ATTENUATION,
+        GL.GL_LINEAR_ATTENUATION,
+        GL.GL_QUADRATIC_ATTENUATION)}, dict())
+
+    # configure lights
+    if params:
+        for mode, value in params:
+            if value is not None:
+                if mode in [GL.GL_AMBIENT, GL.GL_DIFFUSE, GL.GL_SPECULAR,
+                            GL.GL_POSITION]:
+                    lightDesc.params[mode] = (GL.GLfloat * 4)(*value)
+                elif mode == GL.GL_SPOT_DIRECTION:
+                    lightDesc.params[mode] = (GL.GLfloat * 3)(*value)
+                else:
+                    lightDesc.params[mode] = GL.GLfloat(value)
+
+    return lightDesc
+
+
+def useLights(lights, setupOnly=False):
+    """Use specified lights in successive rendering operations. All lights will
+    be transformed using the present modelview matrix.
+
+    Parameters
+    ----------
+    lights : :obj:`List` of :obj:`Light` or None
+        Descriptor of a light source. If None, lighting is disabled.
+    setupOnly : :obj:`bool`, optional
+        Do not enable lighting or lights. Specify True if lighting is being
+        computed via fragment shaders.
+
+    """
+    if lights is not None:
+        if len(lights) > getIntegerv(GL.GL_MAX_LIGHTS):
+            raise IndexError("Number of lights specified > GL_MAX_LIGHTS.")
+
+        GL.glEnable(GL.GL_NORMALIZE)
+
+        for index, light in enumerate(lights):
+            enumLight = GL.GL_LIGHT0 + index
+            # light properties
+            for mode, value in light.params.items():
+                if value is not None:
+                    GL.glLightfv(enumLight, mode, value)
+
+            if not setupOnly:
+                GL.glEnable(enumLight)
+
+        if not setupOnly:
+            GL.glEnable(GL.GL_LIGHTING)
+    else:
+        # disable lights
+        if not setupOnly:
+            for enumLight in range(getIntegerv(GL.GL_MAX_LIGHTS)):
+                GL.glDisable(GL.GL_LIGHT0 + enumLight)
+
+            GL.glDisable(GL.GL_NORMALIZE)
+            GL.glDisable(GL.GL_LIGHTING)
+
+
+def setAmbientLight(color):
+    """Set the global ambient lighting for the scene when lighting is enabled.
+    This is equivalent to GL.glLightModelfv(GL.GL_LIGHT_MODEL_AMBIENT, color)
+    and does not contribute to the GL_MAX_LIGHTS limit.
+
+    Parameters
+    ----------
+    color : :obj:`tuple`
+        Ambient lighting RGBA intensity for the whole scene.
+
+    Notes
+    -----
+    If unset, the default value is (0.2, 0.2, 0.2, 1.0) when GL_LIGHTING is
+    enabled.
+
+    """
+    GL.glLightModelfv(GL.GL_LIGHT_MODEL_AMBIENT, (GL.GLfloat * 4)(*color))
